- Refactor selendroid-standlone to be able to create a server in a remote host instead of always falling back to the localhost.  - Abstract Emulator that implements both AndroidEmulator and AndroidDevice, so other people can have their own emulator version.  - Bug fixes    - Instead of accessing directly the targetPlatform variable, use the get method instead.
