+++ selendroid-standalone/src/main/java/io/selendroid/standalone/server/model/impl/InitAndroidDevicesStrategy.java
@@ -1,0 +1,8 @@
+package io.selendroid.standalone.server.model.impl;
+
+import io.selendroid.standalone.server.model.SelendroidStandaloneDriver;
+
+public interface InitAndroidDevicesStrategy {
+
+    public InitAndroidDevicesConfig getInitAndroidDevicesConfig(SelendroidStandaloneDriver driver);
+}
