+++ selendroid-standalone/src/main/java/io/selendroid/standalone/server/util/HttpClientUtil.java
@@ -41,12 +41,12 @@ public class HttpClientUtil {
   }
 
   public static HttpResponse executeRequestWithPayload(String uri, int port, HttpMethod method,
-      String payload) throws Exception {
+      String payload, String hostname) throws Exception {
     BasicHttpEntityEnclosingRequest request =
         new BasicHttpEntityEnclosingRequest(method.name(), uri);
     request.setEntity(new StringEntity(payload, "UTF-8"));
 
-    return getHttpClient().execute(new HttpHost("localhost", port), request);
+    return getHttpClient().execute(new HttpHost(hostname, port), request);
   }
 
   public static JSONObject parseJsonResponse(HttpResponse response) throws Exception {
