+++ selendroid-standalone/src/test/java/io/selendroid/standalone/server/model/SelendroidStandaloneDriverTest.java
@@ -28,8 +28,6 @@ import io.selendroid.server.common.exceptions.SelendroidException;
 import io.selendroid.standalone.SelendroidConfiguration;
 import io.selendroid.standalone.android.AndroidApp;
 import io.selendroid.standalone.exceptions.DeviceStoreException;
-import io.selendroid.standalone.server.model.DeviceStore;
-import io.selendroid.standalone.server.model.SelendroidStandaloneDriver;
 import io.selendroid.standalone.server.support.DeviceForTest;
 import io.selendroid.standalone.server.support.TestSessionListener;
 
@@ -51,6 +49,25 @@ public class SelendroidStandaloneDriverTest {
   public static final String INVALID_APK_FILE =
       "src/test/resources/selendroid-test-app-invalid.apk";
   private static final Integer EMULATOR_PORT = 5560;
+
+  @Test
+  public void shouldCreateNewTestSession() throws Exception {
+    SelendroidConfiguration conf = new SelendroidConfiguration();
+    conf.addSupportedApp(new File(APK_FILE).getAbsolutePath());
+
+    SelendroidCapabilities caps = new SelendroidCapabilities();
+    caps.setAut(TEST_APP_ID);
+
+    createTestSession(conf, caps);
+  }
+
+  @Test
+  public void shouldCreateNewTestSessionIfNoApkPassed() throws Exception {
+    SelendroidCapabilities caps = new SelendroidCapabilities();
+    caps.setAut(TEST_APP_ID);
+    caps.setLaunchActivity(TEST_APP_LAUNCH_ACTIVITY);
+    createTestSession(new SelendroidConfiguration(), caps);
+  }
 
   @Test
   public void shouldInitDriver() throws Exception {
@@ -103,30 +120,12 @@ public class SelendroidStandaloneDriverTest {
         apps.containsKey(TEST_APP_ID));
   }
 
-  @Test
-  public void shouldCreateNewTestSession() throws Exception {
-    SelendroidConfiguration conf = new SelendroidConfiguration();
-    conf.addSupportedApp(new File(APK_FILE).getAbsolutePath());
-
-    SelendroidCapabilities caps = new SelendroidCapabilities();
-    caps.setAut(TEST_APP_ID);
-
-    createTestSession(conf, caps);
-  }
-
-  @Test
-  public void shouldCreateNewTestSessionIfNoApkPassed() throws Exception {
-    SelendroidCapabilities caps = new SelendroidCapabilities();
-    caps.setAut(TEST_APP_ID);
-    caps.setLaunchActivity(TEST_APP_LAUNCH_ACTIVITY);
-    createTestSession(new SelendroidConfiguration(), caps);
-  }
 
   private void createTestSession(SelendroidConfiguration conf, SelendroidCapabilities caps) throws Exception{
     // Setting up driver with test app and device stub
     SelendroidStandaloneDriver driver = getSelendroidStandaloneDriver();
     driver.initApplicationsUnderTest(conf);
-    DeviceStore store = new DeviceStore(EMULATOR_PORT, getDeviceManager());
+    DefaultDeviceStore store = new DefaultDeviceStore(EMULATOR_PORT, getDeviceManager());
 
     DeviceForTest emulator = new DeviceForTest(DeviceTargetPlatform.ANDROID16);
     Random random = new Random();
@@ -172,7 +171,7 @@ public class SelendroidStandaloneDriverTest {
     conf.setServerStartRetries(configuredRetries);
     driver.initApplicationsUnderTest(conf);
 
-    DeviceStore deviceStore = swapWithFailingDeviceDriver(driver);
+    DefaultDeviceStore deviceStore = swapWithFailingDeviceDriver(driver);
 
     try {
         driver.createNewTestSession(createCapabilities());
@@ -192,7 +191,7 @@ public class SelendroidStandaloneDriverTest {
     conf.setServerStartRetries(0);
     driver.initApplicationsUnderTest(conf);
 
-    DeviceStore deviceStore = swapWithFailingDeviceDriver(driver);
+    DefaultDeviceStore deviceStore = swapWithFailingDeviceDriver(driver);
 
     try {
         driver.createNewTestSession(createCapabilities());
@@ -203,9 +202,9 @@ public class SelendroidStandaloneDriverTest {
       .findAndroidDevice(any(SelendroidCapabilities.class));
   }
 
-  private DeviceStore swapWithFailingDeviceDriver(SelendroidStandaloneDriver driver) throws Exception {
-    //count the amount of calls to DeviceStore to check how many times it was retried
-    DeviceStore deviceStore = mock(DeviceStore.class);
+  private DefaultDeviceStore swapWithFailingDeviceDriver(SelendroidStandaloneDriver driver) throws Exception {
+    //count the amount of calls to DefaultDeviceStore to check how many times it was retried
+    DefaultDeviceStore deviceStore = mock(DefaultDeviceStore.class);
 
     // throw to simulate failure, triggering a retry
     when(deviceStore.findAndroidDevice(any(SelendroidCapabilities.class)))
